+++ src/contour/Config.h
@@ -147,6 +147,7 @@ void loadConfigFromFile(Config& _config, FileSystem::path const& _fileName);
 Config loadConfigFromFile(FileSystem::path const& _fileName);
 Config loadConfig();
 
+std::string createDefaultConfig();
 std::error_code createDefaultConfig(FileSystem::path const& _path);
 std::string defaultConfigFilePath();