+++ src/contour/ContourApp.h
@@ -33,6 +33,7 @@ class ContourApp : public crispy::App
     int parserTableAction();
     int profileAction();
     int terminfoAction();
+    int configAction();
 };
 
 }